requirements.txt
Flask
Flask-Migrate
Flask-SQLAlchemy
Flask-Script
